Conflicts:
	libraries/joomla/form/fields/calendar.php
	tests/unit/suites/libraries/joomla/form/fields/JFormFieldCalendarTest.php

+++ administrator/components/com_menus/models/menu.php
@@ -3,7 +3,7 @@
  * @package     Joomla.Administrator
  * @subpackage  com_menus
  *
- * @copyright   Copyright (C) 2005 - 2014 Open Source Matters, Inc. All rights reserved.
+ * @copyright   Copyright (C) 2005 - 2016 Open Source Matters, Inc. All rights reserved.
  * @license     GNU General Public License version 2 or later; see LICENSE.txt
  */
 
@@ -115,7 +115,6 @@ class MenusModelMenu extends JModelForm
 	public function &getItem($itemId = null)
 	{
 		$itemId = (!empty($itemId)) ? $itemId : (int) $this->getState('menu.id');
-		$false = false;
 
 		// Get a menu item row instance.
 		$table = $this->getTable();
@@ -128,11 +127,11 @@ class MenusModelMenu extends JModelForm
 		{
 			$this->setError($table->getError());
 
-			return $false;
+			return false;
 		}
 
 		$properties = $table->getProperties(1);
-		$value = JArrayHelper::toObject($properties, 'JObject');
+		$value      = JArrayHelper::toObject($properties, 'JObject');
 
 		return $value;
 	}
@@ -274,20 +273,23 @@ class MenusModelMenu extends JModelForm
 		// Iterate the items to delete each one.
 		foreach ($itemIds as $itemId)
 		{
-			// Trigger the before delete event.
-			$result = $dispatcher->trigger('onContentBeforeDelete', array($this->_context, $table));
-
-			if (in_array(false, $result, true) || !$table->delete($itemId))
+			if ($table->load($itemId))
 			{
-				$this->setError($table->getError());
+				// Trigger the before delete event.
+				$result = $dispatcher->trigger('onContentBeforeDelete', array($this->_context, $table));
 
-				return false;
+				if (in_array(false, $result, true) || !$table->delete($itemId))
+				{
+					$this->setError($table->getError());
+
+					return false;
+				}
+
+				// Trigger the after delete event.
+				$dispatcher->trigger('onContentAfterDelete', array($this->_context, $table));
+
+				// TODO: Delete the menu associations - Menu items and Modules
 			}
-
-			// Trigger the after delete event.
-			$dispatcher->trigger('onContentAfterDelete', array($this->_context, $table));
-
-			// TODO: Delete the menu associations - Menu items and Modules
 		}
 
 		// Clean the cache
